requirements.txt
Flask==2.3.3
pytest==7.4.2
pytest-flask==1.3.0
mysql-connector-python==9.1.0
python-dotenv==1.0.1
